Introduced LevelBuilder
First steps to replace existing DungeonBuilder behaviour.
Instead of building a randomised path for the avatar with enemies and rooms placed along it, we start with a grid of connected regions to place rooms in.
Each region maintains a list of connections, with doorways placed between rooms.
If the last room has no connected regions large enough to place a new room, we invert the list and continue from the first room to search for more suitable regions.